Remove duplicate logic and clean up the code.
