--- src/base/hardware_topology.hpp
@@ -1,65 +1,0 @@
-/*
- *
- * Copyright 2016 CUBRID Corporation
- *
- *  Licensed under the Apache License, Version 2.0 (the "License");
- *  you may not use this file except in compliance with the License.
- *  You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- *
- */
-
-/*
- * hardware_topology.hpp
- */
-
-#ifndef _HARDWARE_TOPOLOGY_HPP_
-#define _HARDWARE_TOPOLOGY_HPP_
-
-#ident "$Id$"
-
-#include <hwloc.h>
-#include <vector>
-#include <string>
-#include <cstddef>
-
-namespace cubbase
-{
-  class hardware_topology
-  {
-    public:
-      hardware_topology ();
-      ~hardware_topology ();
-
-      hardware_topology (const hardware_topology &) = delete;
-      hardware_topology &operator= (const hardware_topology &) = delete;
-
-      hardware_topology (hardware_topology &&) = delete;
-      hardware_topology &operator= (hardware_topology &&) = delete;
-
-      void pin_core (int core);
-      std::vector<int> &get_cores ();
-
-      void load_cpu (int limit);
-      void map_nic_to_core ();
-
-    private:
-      hwloc_topology_t m_topology;
-      std::vector<std::vector<int>> m_cores;
-      std::vector<int> m_selected;
-
-      std::string execute_command (const char *cmd);
-      bool set_nic_channels (std::string &ifname, unsigned int combined);
-  };
-
-  extern hardware_topology topology;
-}
-
-#endif
