+++ src/base/system_parameter.c
@@ -4245,29 +4245,6 @@ SYSPRM_PARAM prm_Def[] = {
    (char *) NULL,
    (DUP_PRM_FUNC) NULL,
    (DUP_PRM_FUNC) NULL},
-  {PRM_ID_THREAD_CONNECTION_POOLING,
-   PRM_NAME_THREAD_CONNECTION_POOLING,
-   (PRM_FOR_SERVER),
-   PRM_BOOLEAN,
-   PRM_CLEAR_DYNAMIC_FLAG,
-   {false, {.b = true}},
-   {false, {.b = true}},
-   NULL_SYSPRM_PARAM_VALUE, NULL_SYSPRM_PARAM_VALUE,
-   (char *) NULL,
-   (DUP_PRM_FUNC) NULL,
-   (DUP_PRM_FUNC) NULL},
-  {PRM_ID_THREAD_CONNECTION_TIMEOUT_SECONDS,
-   PRM_NAME_THREAD_CONNECTION_TIMEOUT_SECONDS,
-   (PRM_FOR_SERVER),
-   PRM_INTEGER,
-   PRM_CLEAR_DYNAMIC_FLAG,
-   {false, {.i = 300}},
-   {false, {.i = 300}},
-   {false, {.i = 60 * 60 /* one hour */ }},
-   {false, {.i = -1 /* infinite */ }},
-   (char *) NULL,
-   (DUP_PRM_FUNC) NULL,
-   (DUP_PRM_FUNC) NULL},
   {PRM_ID_THREAD_WORKER_POOLING,
    PRM_NAME_THREAD_WORKER_POOLING,
    (PRM_FOR_SERVER),
@@ -5210,8 +5187,8 @@ SYSPRM_PARAM prm_Def[] = {
    PRM_INTEGER,
    PRM_CLEAR_DYNAMIC_FLAG,
 #if defined (SERVER_MODE)
-   {false, {.i = (int) cubthread::system_core_count ()}},
-   {false, {.i = (int) cubthread::system_core_count ()}},
+   {false, {.i = (int) cubthread::system_core_count () / 2}},
+   {false, {.i = (int) cubthread::system_core_count () / 2}},
    {false, {.i = (int) cubthread::system_core_count ()}},
 #else
    {false, {.i = 2}},
