+++ src/base/system_parameter.h
@@ -413,8 +413,6 @@ enum param_id
   PRM_ID_THREAD_LOGGING_FLAG,
   PRM_ID_LOG_QUERY_LISTS,
 
-  PRM_ID_THREAD_CONNECTION_POOLING,
-  PRM_ID_THREAD_CONNECTION_TIMEOUT_SECONDS,
   PRM_ID_THREAD_WORKER_POOLING,
   PRM_ID_THREAD_WORKER_TIMEOUT_SECONDS,
