+++ src/connection/coordinator.hpp
@@ -23,16 +23,16 @@
 #ifndef _COORDINATOR_HPP_
 #define _COORDINATOR_HPP_
 
-#include "epoll.hpp"
-#include "connection_context.hpp"
-#include "controller.hpp"
-#include "tbb/concurrent_queue.h"
-
 #include <thread>
 #include <vector>
 #include <utility>
 #include <unordered_set>
 #include <unordered_map>
+
+#include "tbb/concurrent_queue.h"
+#include "epoll.hpp"
+#include "connection_context.hpp"
+#include "controller.hpp"
 
 namespace cubconn::connection
 {
